requirements.txt
aiohttp==3.7.3
regex==2021.11.10
python-dateutil==2.8.2
aiodns==3.0.0
dnspython==2.1.0
PyNaCl==1.4.0
psutil==5.8.0
parsedatetime==2.6
numpy==1.21.4
Pillow==8.4.0
git+https://github.com/DisnakeDev/disnake
disnake-jishaku==2.6.5
disnake-ext-menus==0.0.3
pymongo==3.12.1
umongo[motor]
pytz
dateparser